Fixed Issue 3484741: Listing by cost is done alphabetically, not numerically
*****ssw.gui*****
1)dlgOpen.java
Added a NumberRenderer to the cost column of the open dialog. This renderer will take the double and format like it used to be formatted in tbTotalWarfareView and tbChatInformation. This occurs in setupList as well as the event handler for switching the table views.

2)NumberRenderer.java
Added a new TableCellRenderer based on DefaultTableCellRenderer class, with the exception that it will attempt to parse the object into a double then format the double accordingly.